# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gözlük Bayi Yönetimi API (MVB)",
    "version": "1.3.0",
    "description": "Bayi yönetimi, sipariş takibi, çok kiracılı kullanıcı yönetimi ve dosya depolama (Storage) işlemleri için backend API dökümantasyonu."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "API Sunucusu"
    },
    {
      "url": "http://localhost:3000/l",
      "description": "Kısa Link Sunucusu"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/dealers": {
      "get": {
        "tags": [
          "Dealers"
        ],
        "summary": "Tüm bayileri listele",
        "responses": {
          "200": {
            "description": "Bayi listesi."
          }
        }
      },
      "post": {
        "tags": [
          "Dealers"
        ],
        "summary": "Yeni bayi oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bayi oluşturuldu."
          },
          "401": {
            "description": "Yetkisiz erişim."
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Tüm siparişleri listele",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sipariş listesi."
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Yeni sipariş oluştur (Konfigüratörden)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dealerId",
                  "configuration"
                ],
                "properties": {
                  "dealerId": {
                    "type": "integer"
                  },
                  "customerName": {
                    "type": "string"
                  },
                  "configuration": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sipariş oluşturuldu."
          }
        }
      }
    },
...
    "/appointments": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Tüm randevuları listele",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Randevu listesi."
          }
        }
      },
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Yeni randevu oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dealerId",
                  "customerName",
                  "appointmentDate"
                ],
                "properties": {
                  "dealerId": {
                    "type": "integer"
                  },
                  "customerName": {
                    "type": "string"
                  },
                  "appointmentDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "Eye Exam",
                      "Styling",
                      "Adjustment",
                      "Other"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Randevu başarıyla oluşturuldu."
          }
        }
      }
    },
...
  }
}